refactor the interface and fixed the styles
