+++ examples/embd-input/embd-input-test.cpp
@@ -1,4 +1,4 @@
-#include "embd_input.h"
+#include "embd-input.h"
 #include <stdlib.h>
 #include <random>
 #include <string.h>
@@ -7,8 +7,11 @@ int main(int argc, char** argv) {
 
     auto mymodel = create_mymodel(argc, argv);
     int N = 10;
+    int max_tgt_len = 500;
     int n_embd = llama_n_embd(mymodel->ctx);
-    float* data = new float[N*n_embd];
+
+    // add random float embd to test evaluation
+    float * data = new float[N*n_embd];
     std::default_random_engine e;
     std::uniform_real_distribution<float>  u(0,1);
     for (int i=0;i<N*n_embd;i++) {
@@ -16,19 +19,15 @@ int main(int argc, char** argv) {
     }
 
     eval_string(mymodel, "user: what is the color of the flag of UN?");
-    // printf("eval float");
     eval_float(mymodel, data, N);
     eval_string(mymodel, "assistant:");
-    // printf("eval float end\n");
     eval_string(mymodel, mymodel->params.prompt.c_str());
     const char* tmp;
-    for (int i=0;i < 500; i++) {
-        // int id = sampling_id(mymodel);
+    for (int i=0; i<max_tgt_len; i++) {
         tmp = sampling(mymodel);
         if (strcmp(tmp, "</s>")==0) break;
-        printf("%s", tmp); // llama_token_to_str(mymodel->ctx, id));
+        printf("%s", tmp);
         fflush(stdout);
-        // eval_id(mymodel, id);
     }
     printf("\n");
     free_mymodel(mymodel);
