+++ examples/embd-input/embd-input.h
